Import hTC zip | 3.10.49 | g6f6e344

+++ drivers/media/platform/msm/camera_v2/ispif/msm_ispif.c
@@ -125,11 +125,11 @@ static int msm_ispif_reset_hw(struct ispif_device *ispif)
 			pr_err("%s: cannot enable clock, error = %d",
 				__func__, rc);
 		} else {
-			/* This is set when device is 8x26 */
+			
 			ispif->clk_idx = 2;
 		}
 	} else {
-		/* This is set when device is 8974 */
+		
 		ispif->clk_idx = 1;
 	}
 
@@ -137,7 +137,7 @@ static int msm_ispif_reset_hw(struct ispif_device *ispif)
 	if (ispif->hw_num_isps > 1)
 		init_completion(&ispif->reset_complete[VFE1]);
 
-	/* initiate reset of ISPIF */
+	
 	msm_camera_io_w(ISPIF_RST_CMD_MASK,
 				ispif->base + ISPIF_RST_CMD_ADDR);
 	if (ispif->hw_num_isps > 1)
@@ -259,7 +259,7 @@ static int msm_ispif_clk_ahb_enable(struct ispif_device *ispif, int enable)
 	int rc = 0;
 
 	if (ispif->csid_version < CSID_VERSION_V30) {
-		/* Older ISPIF versiond don't need ahb clokc */
+		
 		return 0;
 	}
 
@@ -358,23 +358,23 @@ static void msm_ispif_sel_csid_core(struct ispif_device *ispif,
 	switch (intftype) {
 	case PIX0:
 		data &= ~(BIT(1) | BIT(0));
-		data |= csid;
+		data |= (uint32_t) csid;
 		break;
 	case RDI0:
 		data &= ~(BIT(5) | BIT(4));
-		data |= (csid << 4);
+		data |= ((uint32_t) csid) << 4;
 		break;
 	case PIX1:
 		data &= ~(BIT(9) | BIT(8));
-		data |= (csid << 8);
+		data |= ((uint32_t) csid) << 8;
 		break;
 	case RDI1:
 		data &= ~(BIT(13) | BIT(12));
-		data |= (csid << 12);
+		data |= ((uint32_t) csid) << 12;
 		break;
 	case RDI2:
 		data &= ~(BIT(21) | BIT(20));
-		data |= (csid << 20);
+		data |= ((uint32_t) csid) << 20;
 		break;
 	}
 
@@ -450,9 +450,9 @@ static void msm_ispif_enable_intf_cids(struct ispif_device *ispif,
 
 	data = msm_camera_io_r(ispif->base + intf_addr);
 	if (enable)
-		data |= cid_mask;
+		data |=  (uint32_t) cid_mask;
 	else
-		data &= ~cid_mask;
+		data &= ~((uint32_t) cid_mask);
 	msm_camera_io_w_mb(data, ispif->base + intf_addr);
 }
 
@@ -704,28 +704,28 @@ static void msm_ispif_intf_cmd(struct ispif_device *ispif, uint32_t cmd_bits,
 			cid = params->entries[i].cids[k];
 			vc = cid / 4;
 			if (intf_type == RDI2) {
-				/* zero out two bits */
+				
 				ispif->applied_intf_cmd[vfe_intf].intf_cmd1 &=
 					~(0x3 << (vc * 2 + 8));
-				/* set cmd bits */
+				
 				ispif->applied_intf_cmd[vfe_intf].intf_cmd1 |=
 					(cmd_bits << (vc * 2 + 8));
 			} else {
-				/* zero 2 bits */
+				
 				ispif->applied_intf_cmd[vfe_intf].intf_cmd &=
 					~(0x3 << (vc * 2 + intf_type * 8));
-				/* set cmd bits */
+				
 				ispif->applied_intf_cmd[vfe_intf].intf_cmd |=
 					(cmd_bits << (vc * 2 + intf_type * 8));
 			}
 		}
-		/* cmd for PIX0, PIX1, RDI0, RDI1 */
+		
 		if (ispif->applied_intf_cmd[vfe_intf].intf_cmd != 0xFFFFFFFF)
 			msm_camera_io_w_mb(
 				ispif->applied_intf_cmd[vfe_intf].intf_cmd,
 				ispif->base + ISPIF_VFE_m_INTF_CMD_0(vfe_intf));
 
-		/* cmd for RDI2 */
+		
 		if (ispif->applied_intf_cmd[vfe_intf].intf_cmd1 != 0xFFFFFFFF)
 			msm_camera_io_w_mb(
 				ispif->applied_intf_cmd[vfe_intf].intf_cmd1,
@@ -756,7 +756,7 @@ static int msm_ispif_stop_immediately(struct ispif_device *ispif,
 	}
 	msm_ispif_intf_cmd(ispif, ISPIF_INTF_CMD_DISABLE_IMMEDIATELY, params);
 
-	/* after stop the interface we need to unmask the CID enable bits */
+	
 	for (i = 0; i < params->num; i++) {
 		cid_mask = msm_ispif_get_cids_mask_from_cfg(
 			&params->entries[i]);
@@ -838,18 +838,18 @@ static int msm_ispif_restart_frame_boundary(struct ispif_device *ispif,
 			pr_err("%s: cannot enable clock, error = %d",
 				__func__, rc);
 		} else {
-			/* This is set when device is 8x26 */
+			
 			ispif->clk_idx = 2;
 		}
 	} else {
-		/* This is set when device is 8974 */
+		
 		ispif->clk_idx = 1;
 	}
 
 	if (vfe_mask & (1 << VFE0)) {
 		init_completion(&ispif->reset_complete[VFE0]);
 		pr_err("%s Init completion VFE0\n", __func__);
-			/* initiate reset of ISPIF */
+			
 		msm_camera_io_w(0x00001FF9,
 				ispif->base + ISPIF_RST_CMD_ADDR);
 	}
@@ -1041,7 +1041,7 @@ static int msm_ispif_stop_frame_boundary(struct ispif_device *ispif,
 		if (rc < 0)
 			goto end;
 
-		/* disable CIDs in CID_MASK register */
+		
 		msm_ispif_enable_intf_cids(ispif, params->entries[i].intftype,
 			cid_mask, vfe_intf, 0);
 	}
@@ -1189,7 +1189,7 @@ static int msm_ispif_init(struct ispif_device *ispif,
 		return rc;
 	}
 
-	/* can we set to zero? */
+	
 	ispif->applied_intf_cmd[VFE0].intf_cmd  = 0xFFFFFFFF;
 	ispif->applied_intf_cmd[VFE0].intf_cmd1 = 0xFFFFFFFF;
 	ispif->applied_intf_cmd[VFE1].intf_cmd  = 0xFFFFFFFF;
@@ -1267,7 +1267,7 @@ static void msm_ispif_release(struct ispif_device *ispif)
 		return;
 	}
 
-	/* make sure no streaming going on */
+	
 	msm_ispif_reset(ispif);
 
 	msm_ispif_clk_ahb_enable(ispif, 0);
@@ -1294,7 +1294,7 @@ static long msm_ispif_cmd(struct v4l2_subdev *sd, void *arg)
 	mutex_lock(&ispif->mutex);
 	switch (pcdata->cfg_type) {
 	case ISPIF_ENABLE_REG_DUMP:
-		ispif->enb_dump_reg = pcdata->reg_dump; /* save dump config */
+		ispif->enb_dump_reg = pcdata->reg_dump; 
 		break;
 	case ISPIF_INIT:
 		rc = msm_ispif_init(ispif, pcdata->csid_version);
@@ -1383,7 +1383,7 @@ static int ispif_open_node(struct v4l2_subdev *sd, struct v4l2_subdev_fh *fh)
 	struct ispif_device *ispif = v4l2_get_subdevdata(sd);
 
 	mutex_lock(&ispif->mutex);
-	/* mem remap is done in init when the clock is on */
+	
 	ispif->open_cnt++;
 	mutex_unlock(&ispif->mutex);
 	return 0;
@@ -1444,9 +1444,9 @@ static int ispif_probe(struct platform_device *pdev)
 		rc = of_property_read_u32((&pdev->dev)->of_node,
 		"qcom,num-isps", &ispif->hw_num_isps);
 		if (rc)
-			/* backward compatibility */
+			
 			ispif->hw_num_isps = 1;
-		/* not an error condition */
+		
 		rc = 0;
 	}
